Removing default logger. basic logger accessible. better protocol naming
